+++ src/main/java/ba/unsa/etf/si/mainserver/models/auth/User.java
@@ -1,5 +1,6 @@
-package ba.unsa.etf.si.mainserver.models;
+package ba.unsa.etf.si.mainserver.models.auth;
 
+import ba.unsa.etf.si.mainserver.models.AuditModel;
 import lombok.*;
 
 import javax.persistence.*;
