+++ src/main/java/ba/unsa/etf/si/mainserver/repositories/auth/RoleRepository.java
@@ -1,11 +1,10 @@
-package ba.unsa.etf.si.mainserver.repositories;
+package ba.unsa.etf.si.mainserver.repositories.auth;
 
-import ba.unsa.etf.si.mainserver.models.Role;
-import ba.unsa.etf.si.mainserver.models.RoleName;
+import ba.unsa.etf.si.mainserver.models.auth.Role;
+import ba.unsa.etf.si.mainserver.models.auth.RoleName;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
