+++ src/main/java/ba/unsa/etf/si/mainserver/requests/auth/RegistrationRequest.java
@@ -1,5 +1,6 @@
-package ba.unsa.etf.si.mainserver.responses;
+package ba.unsa.etf.si.mainserver.requests.auth;
 
+import ba.unsa.etf.si.mainserver.responses.auth.RoleResponse;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,8 +10,7 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class RegistrationResponse {
-    private Long id;
+public class RegistrationRequest {
     private String username;
     private String password;
     private String email;
