+++ src/main/java/ba/unsa/etf/si/mainserver/services/UserService.java
@@ -2,13 +2,13 @@ package ba.unsa.etf.si.mainserver.services;
 
 import ba.unsa.etf.si.mainserver.exceptions.AppException;
 import ba.unsa.etf.si.mainserver.exceptions.UnauthorizedException;
-import ba.unsa.etf.si.mainserver.models.Role;
-import ba.unsa.etf.si.mainserver.models.RoleName;
-import ba.unsa.etf.si.mainserver.models.User;
-import ba.unsa.etf.si.mainserver.repositories.RoleRepository;
-import ba.unsa.etf.si.mainserver.repositories.UserRepository;
-import ba.unsa.etf.si.mainserver.requests.LoginRequest;
-import ba.unsa.etf.si.mainserver.requests.RegistrationRequest;
+import ba.unsa.etf.si.mainserver.models.auth.Role;
+import ba.unsa.etf.si.mainserver.models.auth.RoleName;
+import ba.unsa.etf.si.mainserver.models.auth.User;
+import ba.unsa.etf.si.mainserver.repositories.auth.RoleRepository;
+import ba.unsa.etf.si.mainserver.repositories.auth.UserRepository;
+import ba.unsa.etf.si.mainserver.requests.auth.LoginRequest;
+import ba.unsa.etf.si.mainserver.requests.auth.RegistrationRequest;
 import ba.unsa.etf.si.mainserver.security.JwtTokenProvider;
 import ba.unsa.etf.si.mainserver.security.UserCreationPermissions;
 import ba.unsa.etf.si.mainserver.security.UserPrincipal;
